BTreeMap/BTreeSet: separate off code supporting tests
